+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added 
+
+- Suppress xarray dependency, change of API to use numpy array. [#42]
+
 ## 1.1.0 (January 2023)
 
 ### Added